More fine-grained notifications to watchpoint
This breaking change changes the arguments to the watch functions on
the `watchpoint` var. The third argument is now one of `:starting`,
`:started`, `:stopping` and `:stopped`. The fourth argument is now
always the State object.

This allows for more fine-grained control over the state (changes).
For example, logging can be more thorough. Another example is that a
dependency graph can be deduced.